+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ##  Unreleased
 
+### Changed
+
+- Allow the request handler before method to return an early response
+
 ## 0.3.0 (2018-09-04)
 
 ### Added